Unify Custom Gas Components (#1860)
* Remove redundant TransactionEdit component, move logic to TransactionEditor

* add AnimatedCustomGas

* Move all animation logic into Animated Transaction Modal

* TransactionReviewFeeCard: change toggleCustomGasModal to edit

* toggleCustomGasModal --> edit

* AnimatedTransactionModal: Make animations function from wallet send flow, custom gas modal renders properly

* Confirm: use CustomGas from UI, visually in working state but not yet functionally

* Confirm: setting custom gas from selectors or input works

* Bug fix: can't tap edit before gas estimation fully complete

* bug fix: gas not showing after selection

* Confirm: rafactor, move class functions

* remove redundant function

* ApproveTransactionReview: new components. Takes logic from ApproveView/Approve. Remove dead code

* Approve: break out logic into ApproveTransactionReview, plug with AnimatedTransactionModal, use UI/CustomGas, remove dead code

* ApprveTransactionReview: add componentDidUpdate to update gas

* formatting

* ApproveTransactionView: use ActionView for buttons

* ESLINT: remove dumb componentDidUpdate no setState rule

* Fix paddings

* change confirm to approve

* add new test + update snapshots

* REMOVE SendFlow/CustomGas RIP

* use component variables for animated values instead of state

* use optional chaining on react children

* Customas: remove '' as possible value to checksummedFrom

* AnimatedTransactionModal: dedupe code, move all logic into render

* Apply changes from #1847 inside ApproveTransactionReview

* update snapshot

* rafactor gas speed selection, put in parent state (Confirm) to pass down as props

* Only apply buttonStyle animation for iOS

* Leave space for the gas error

* Add margin to the bottom as well

* Add no-did-update-set-state rule back and disable where we're doing it instead
