+++ app/components/UI/AnimatedTransactionModal/index.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import AnimatedTransactionModal from './';
+
+describe('AnimatedTransactionModal', () => {
+	it('should render correctly', () => {
+		const wrapper = shallow(<AnimatedTransactionModal />);
+		expect(wrapper).toMatchSnapshot();
+	});
+});
