+++ app/components/UI/TransactionEditor/index.js
@@ -1,12 +1,12 @@
 import React, { PureComponent } from 'react';
-import { StyleSheet, Animated, Easing } from 'react-native';
+import { StyleSheet } from 'react-native';
 import PropTypes from 'prop-types';
-import { colors } from '../../../styles/common';
 import ConfirmSend from '../../Views/SendFlow/Confirm';
+import AnimatedTransactionModal from '../AnimatedTransactionModal';
 import TransactionReview from '../TransactionReview';
-import TransactionEdit from '../TransactionEdit';
-import { isBN, hexToBN, toBN, isDecimal } from '../../../util/number';
-import { isValidAddress, toChecksumAddress, BN } from 'ethereumjs-util';
+import CustomGas from '../CustomGas';
+import { isBN, hexToBN, toBN, isDecimal, fromWei } from '../../../util/number';
+import { isValidAddress, toChecksumAddress, BN, addHexPrefix } from 'ethereumjs-util';
 import { strings } from '../../../../locales/i18n';
 import { connect } from 'react-redux';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
@@ -19,27 +19,11 @@ import contractMap from 'eth-contract-metadata';
 import PaymentChannelsClient from '../../../core/PaymentChannelsClient';
 import { safeToChecksumAddress } from '../../../util/address';
 import TransactionTypes from '../../../core/TransactionTypes';
-import Device from '../../../util/Device';
 
 const EDIT = 'edit';
 const REVIEW = 'review';
 
 const styles = StyleSheet.create({
-	root: {
-		backgroundColor: colors.white,
-		minHeight: 200,
-		borderTopLeftRadius: 20,
-		borderTopRightRadius: 20,
-		paddingBottom: Device.isIphoneX() ? 24 : 0
-	},
-	transactionEdit: {
-		position: 'absolute',
-		width: '100%',
-		height: '100%'
-	},
-	transactionReview: {
-		paddingTop: 24
-	},
 	keyboardAwareWrapper: {
 		flex: 1,
 		justifyContent: 'flex-end'
@@ -56,10 +40,6 @@ class TransactionEditor extends PureComponent {
 		 */
 		accounts: PropTypes.object,
 		/**
-		 * react-navigation object used for switching between screens
-		 */
-		navigation: PropTypes.object,
-		/**
 		 * A string representing the network name
 		 */
 		networkType: PropTypes.string,
@@ -72,7 +52,7 @@ class TransactionEditor extends PureComponent {
 		 */
 		onCancel: PropTypes.func,
 		/**
-		 * Callback triggered when this transaction is cancelled
+		 * Callback triggered when this transaction is confirmed
 		 */
 		onConfirm: PropTypes.func,
 		/**
@@ -118,28 +98,31 @@ class TransactionEditor extends PureComponent {
 		ensRecipient: undefined,
 		basicGasEstimates: {},
 		ready: false,
-		advancedCustomGas: false,
-		modalValue: new Animated.Value(1),
-		reviewToEditValue: new Animated.Value(0),
-		reviewToDataValue: new Animated.Value(0),
-		editToAdvancedValue: new Animated.Value(0),
-		width: Device.getDeviceWidth(),
-		rootHeight: null,
-		customGasHeight: null,
-		transactionReviewDataHeight: null,
-		hideGasSelectors: false,
-		hideData: true,
-		toAdvancedFrom: 'edit'
-	};
-
-	xTranslationMappings = {
-		reviewToEdit: this.state.reviewToEditValue,
-		editToAdvanced: this.state.editToAdvancedValue,
-		reviewToData: this.state.reviewToDataValue
+		data: undefined,
+		amountError: '',
+		gasError: '',
+		toAddressError: ''
 	};
 
 	componentDidMount = async () => {
+		const { transaction } = this.props;
 		await this.handleFetchBasicEstimates();
+		if (transaction && transaction.value) {
+			this.handleUpdateAmount(transaction.value, true);
+		}
+		if (transaction && transaction.assetType === 'ETH') {
+			this.handleUpdateReadableValue(fromWei(transaction.value));
+		}
+		if (transaction && transaction.data) {
+			this.setState({ data: transaction.data });
+		}
+	};
+
+	componentDidUpdate = prevProps => {
+		const { transaction } = this.props;
+		if (transaction.data !== prevProps.transaction.data) {
+			this.handleUpdateData(transaction.data);
+		}
 	};
 
 	/**
@@ -235,6 +218,7 @@ class TransactionEditor extends PureComponent {
 	 */
 	handleUpdateData = async data => {
 		const { gas } = await this.estimateGas({ data });
+		this.setState({ data });
 		this.props.setTransactionObject({ gas: hexToBN(gas), data });
 	};
 
@@ -570,6 +554,31 @@ class TransactionEditor extends PureComponent {
 		return undefined;
 	};
 
+	review = async () => {
+		const { data } = this.state;
+		await this.setState({ toFocused: true });
+		const validated = !(await this.validate());
+		if (validated) {
+			if (data && data.substr(0, 2) !== '0x') {
+				this.handleUpdateData(addHexPrefix(data));
+			}
+		}
+	};
+
+	validate = async () => {
+		const amountError = await this.validateAmount(false);
+		const gasError = this.validateGas();
+		const toAddressError = this.validateToAddress();
+		this.setState({ amountError, gasError, toAddressError });
+		return amountError || gasError || toAddressError;
+	};
+
+	updateGas = async (gas, gasLimit) => {
+		await this.handleGasFeeSelection(gas, gasLimit);
+		const gasError = this.validateGas();
+		this.setState({ gasError });
+	};
+
 	handleNewTxMeta = async ({
 		target_address,
 		chain_id = null, // eslint-disable-line no-unused-vars
@@ -606,187 +615,32 @@ class TransactionEditor extends PureComponent {
 		this.setState({ basicGasEstimates, ready: true });
 	};
 
-	onModeChange = mode => {
-		if (mode === 'edit') {
-			this.setState({ toAdvancedFrom: 'review' });
-			this.animate({
-				modalEndValue: this.state.advancedCustomGas ? this.getAnimatedModalValueForAdvancedCG() : 0,
-				xTranslationName: 'reviewToEdit',
-				xTranslationEndValue: 1
-			});
-		} else {
-			this.animate({
-				modalEndValue: 1,
-				xTranslationName: 'reviewToEdit',
-				xTranslationEndValue: 0
-			});
-		}
-		this.props.onModeChange(mode);
-	};
-
-	toggleAdvancedCustomGas = (toggle = false) => {
-		const { advancedCustomGas } = this.state;
-		this.setState({ advancedCustomGas: toggle ? true : !advancedCustomGas, toAdvancedFrom: 'edit' });
-	};
-
-	hideComponents = (xTranslationName, xTranslationEndValue, animationTime) => {
-		//data view is hidden by default because when we switch from review to edit, since view is nested in review, it also gets transformed. It's shown if it's the animation's destination.
-		if (xTranslationName === 'editToAdvanced') {
-			this.setState({
-				hideGasSelectors: xTranslationEndValue === 1 && animationTime === 'end'
-			});
-		}
-		if (xTranslationName === 'reviewToData') {
-			this.setState({
-				hideData: xTranslationEndValue === 0 && animationTime === 'end'
-			});
-		}
-	};
-
-	animate = ({ modalEndValue, xTranslationName, xTranslationEndValue }) => {
-		const { modalValue } = this.state;
-		this.hideComponents(xTranslationName, xTranslationEndValue, 'start');
-		Animated.parallel([
-			Animated.timing(modalValue, {
-				toValue: modalEndValue,
-				duration: 250,
-				easing: Easing.ease,
-				useNativeDriver: true
-			}),
-			Animated.timing(this.xTranslationMappings[xTranslationName], {
-				toValue: xTranslationEndValue,
-				duration: 250,
-				easing: Easing.ease,
-				useNativeDriver: true
-			})
-		]).start(() => {
-			this.hideComponents(xTranslationName, xTranslationEndValue, 'end');
-		});
-	};
-
-	generateTransform = (valueType, outRange) => {
-		const { modalValue, reviewToEditValue, editToAdvancedValue, reviewToDataValue } = this.state;
-		if (valueType === 'modal' || valueType === 'saveButton') {
-			return {
-				transform: [
-					{
-						translateY: modalValue.interpolate({
-							inputRange: [0, valueType === 'saveButton' ? this.getAnimatedModalValueForAdvancedCG() : 1],
-							outputRange: outRange
-						})
-					}
-				]
-			};
-		}
-		let value;
-		if (valueType === 'reviewToEdit') value = reviewToEditValue;
-		else if (valueType === 'editToAdvanced') value = editToAdvancedValue;
-		else if (valueType === 'reviewToData') value = reviewToDataValue;
-		return {
-			transform: [
-				{
-					translateX: value.interpolate({
-						inputRange: [0, 1],
-						outputRange: outRange
-					})
-				}
-			]
-		};
-	};
-
-	getAnimatedModalValueForAdvancedCG = () => {
-		const { rootHeight, customGasHeight } = this.state;
-		//70 is the fixed height + margin of the error message in advanced custom gas. It expands 70 units vertically to accomodate it
-		return 70 / (rootHeight - customGasHeight);
-	};
-
-	saveRootHeight = event => this.setState({ rootHeight: event.nativeEvent.layout.height });
-
-	saveCustomGasHeight = event =>
-		!this.state.customGasHeight && this.setState({ customGasHeight: event.nativeEvent.layout.height });
-
-	saveTransactionReviewDataHeight = event =>
-		!this.state.transactionReviewDataHeight &&
-		this.setState({ transactionReviewDataHeight: event.nativeEvent.layout.height });
-
-	getTransformValue = () => {
-		const { rootHeight, customGasHeight } = this.state;
-		return rootHeight - customGasHeight;
-	};
 	render = () => {
-		const { mode, transactionConfirmed, transaction } = this.props;
-		const {
-			basicGasEstimates,
-			ready,
-			width,
-			advancedCustomGas,
-			hideGasSelectors,
-			modalValue,
-			toAdvancedFrom,
-			hideData,
-			customGasHeight
-		} = this.state;
+		const { mode, transactionConfirmed, transaction, onModeChange } = this.props;
+		const { basicGasEstimates, ready, gasError } = this.state;
 		const paymentChannelTransaction = transaction ? transaction.paymentChannelTransaction : false;
 		return (
 			<React.Fragment>
 				{mode === EDIT && paymentChannelTransaction && <ConfirmSend transaction={transaction} />}
 				{!paymentChannelTransaction && (
 					<KeyboardAwareScrollView contentContainerStyle={styles.keyboardAwareWrapper}>
-						<Animated.View
-							style={[styles.root, this.generateTransform('modal', [this.getTransformValue(), 0])]}
-							onLayout={this.saveRootHeight}
-						>
-							<Animated.View
-								style={[this.generateTransform('reviewToEdit', [0, -width]), styles.transactionReview]}
-							>
-								<TransactionReview
-									onCancel={this.onCancel}
-									onConfirm={this.onConfirm}
-									onModeChange={this.onModeChange}
-									validate={this.validate}
-									ready={ready}
-									transactionConfirmed={transactionConfirmed}
-									generateTransform={this.generateTransform}
-									animate={this.animate}
-									hideData={hideData}
-									saveTransactionReviewDataHeight={this.saveTransactionReviewDataHeight}
-									customGasHeight={customGasHeight}
-								/>
-							</Animated.View>
-							{ready && (
-								<Animated.View
-									style={[styles.transactionEdit, this.generateTransform('reviewToEdit', [width, 0])]}
-								>
-									<TransactionEdit
-										navigation={this.props.navigation}
-										basicGasEstimates={basicGasEstimates}
-										onCancel={this.onCancel}
-										onModeChange={this.onModeChange}
-										handleUpdateAmount={this.handleUpdateAmount}
-										handleUpdateData={this.handleUpdateData}
-										handleUpdateFromAddress={this.handleUpdateFromAddress}
-										handleUpdateToAddress={this.handleUpdateToAddress}
-										handleGasFeeSelection={this.handleGasFeeSelection}
-										validateAmount={this.validateAmount}
-										validateGas={this.validateGas}
-										validateToAddress={this.validateToAddress}
-										handleUpdateAsset={this.handleUpdateAsset}
-										checkForAssetAddress={this.checkForAssetAddress}
-										handleUpdateReadableValue={this.handleUpdateReadableValue}
-										saveCustomGasHeight={this.saveCustomGasHeight}
-										toggleAdvancedCustomGas={this.toggleAdvancedCustomGas}
-										advancedCustomGas={advancedCustomGas}
-										animate={this.animate}
-										generateTransform={this.generateTransform}
-										getAnimatedModalValueForAdvancedCG={this.getAnimatedModalValueForAdvancedCG}
-										hideGasSelectors={hideGasSelectors}
-										mode={mode}
-										modalValue={modalValue}
-										toAdvancedFrom={toAdvancedFrom}
-									/>
-								</Animated.View>
-							)}
-						</Animated.View>
+						<AnimatedTransactionModal onModeChange={onModeChange} ready={ready} review={this.review}>
+							<TransactionReview
+								onCancel={this.onCancel}
+								onConfirm={this.onConfirm}
+								validate={this.validate}
+								ready={ready}
+								transactionConfirmed={transactionConfirmed}
+							/>
+							<CustomGas
+								handleGasFeeSelection={this.updateGas}
+								basicGasEstimates={basicGasEstimates}
+								gas={transaction.gas}
+								gasPrice={transaction.gasPrice}
+								gasError={gasError}
+								mode={mode}
+							/>
+						</AnimatedTransactionModal>
 					</KeyboardAwareScrollView>
 				)}
 			</React.Fragment>
